Prevent packages being downloaded unnecessarily
Version numbers were being converted to floating point values during
YAML parsing which caused cacophony.pkg_installed_from_github to not
think the correct version of a package was installed and therefore
download it again.

All version numbers in states have been forced to strings and
pkg_installed_from_github now asserts that the version number it gets is
a string.

On an up-to-date pi with a good network connection this brings the
state.apply time down from 61.7s to 11.9s.